# Inception fusion operator.