--- a/python/paddle/fluid/tests/unittests/testsuite.py
+++ b/python/paddle/fluid/tests/unittests/testsuite.py
@@ -139,8 +139,9 @@ def append_input_output(block, op_proto, np_list, is_input, dtype):
         var_name = str(var_proto.name)
         if (var_name not in np_list) and var_proto.dispensable:
             continue
-        assert (var_name in np_list) or (var_proto.dispensable), \
-            "Missing {} as input".format(var_name)
+        if is_input:
+            assert (var_name in np_list) or (var_proto.dispensable), \
+                "Missing {} as input".format(var_name)
         if var_proto.duplicable:
             assert isinstance(np_list[var_name], list), \
                 "Duplicable {} should be set as list".format(var_name)