… paddle/fluid/operators/fused/CMakeLists.txt
… paddle/fluid/operators/fused/fusion_conv_inception_op.cu