1.  Inception fusion:

    Speedup:

    run 1000 times:

    ```
    inception_a1:  523.8522 ms ->  252.816 ms
    inception_a2:  579.2627 ms -> 251.684 ms
    inception_a3:  580.5288 ms -> 245.702 ms
    ```

2. Support horizontal layer fusion in conv_fusion_op. 
3. Add conv search strategy for variable-length input.